월 상환가능금액 Long으로 통일

+++ scoula/backend/src/main/java/org/scoula/member/mapper/MemberMapper.java
@@ -20,6 +20,6 @@ public interface MemberMapper {
 
     int updateCreditScore(@Param("userId") Long userId, @Param("creditScore") Integer creditScore);
 
-    int updateMaxMonthlyPayment(@Param("userId") Long userId, @Param("maxMonthlyPayment") Integer maxMonthlyPayment);
+    int updateMaxMonthlyPayment(@Param("userId") Long userId, @Param("maxMonthlyPayment") Long maxMonthlyPayment);
 
 }
